+++ server/src/util/server_root.ts
@@ -1,5 +1,5 @@
 /**
- * Logic for finding the server root directory to locate config, DB and migration files.
+ * Logic for finding the server root directory to locate data and migration files.
  * This is required as working directory might differ across local development, docker and server-cli.
  */
 
@@ -7,7 +7,7 @@ import fs from 'node:fs'
 import path from 'node:path'
 import { Result } from 'better-result'
 
-const MARKERS = ['drizzle', 'config.toml']
+const MARKERS = ['drizzle', 'data']
 
 /**
  * Check if directory contains any marker files